[dtensor] enable foreach operators for adam optimizer
This PR enables basic foreach ops in DTensor for adam optimizer, to improve performance
compare to optimizer using torch.Tensor. Currently by default optimizer
won't do this for tensor subclass, we will need to enable this by
default in DTensor when all ops are covered, or early enable this when
exploring new FSDP, we just need to append DTensor to the optimizer
allow list.

cc awgu

ghstack-source-id: bf89d590477fe43dc8f00220ac9713e3c775a442
Pull Request resolved: https://github.com/pytorch/pytorch/pull/112108