Update how Dynamo decides to graph break on an OpOverloadPacket
Previously, under config.only_allow_pt2_compliant_ops, Dynamo graph
breaks when it see an OpOverloadPacket where any overloads are not
PT2 compliant. This is potentially brittle: if someone (unlikely) adds
a new overload for a custom operator, then this would cause a
previously non-graph-breaking call to the OpOverloadPacket to graph break.

In this PR:
- When the operator is invoked (via call_function), we first check if
it is an OpOverload or an OpOverloadPacket. If it is an OpOverloadPacket
with a single overload, then we check the pt2_compliant_tag.
- If the OpOverloadPacket has multiple overloads, then we attempt to
grab the fake values on the node that we create and then use those to
resolve the overload. Then, we check if the resolved overload has the
pt2_compliant_tag

Test Plan:
- new tests, existing tests

ghstack-source-id: a1d8cd4efa9b3a994b6e7ff575c69c7e82c14c0e
Pull Request resolved: https://github.com/pytorch/pytorch/pull/112200