[inductor] Fix bug handling output_strides in fx graph cache
Summary: The current implementation is not properly attaching output strides to the tracing context when an fx graph is loaded from the cache. That bugs leads to assertion failures like `AssertionError: expected size 3==3, stride 1==9 at dim=1`. This change saves the output strides in the serialized object cached on disk and inserts them into the tracing context whether the graph is loaded from cache or compiled.

Test Plan:
* New unit test using resnet18 (which repros the problem)
* Ran the timm benchmark suite with `--training`

ghstack-source-id: a87856987eeeed9cf2079f531d230c0a8a047670
Pull Request resolved: https://github.com/pytorch/pytorch/pull/112041